+++ packages/api/app/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @modern-js/codesmith-api-app
 
+## 1.0.0-rc.14
+
+### Patch Changes
+
+- 1e2e787: feat: initial
+- Updated dependencies [1e2e787]
+  - @modern-js/codesmith-api-git@1.0.0-rc.14
+  - @modern-js/codesmith-api-handlebars@1.0.0-rc.14
+  - @modern-js/codesmith-api-npm@1.0.0-rc.14
+  - @modern-js/codesmith@1.0.0-rc.14
+  - @modern-js/easy-form-cli@1.0.0-rc.14
+  - @modern-js/inquirer-types@1.0.0-rc.14
+
+## 1.0.0-rc.13
+
+### Patch Changes
+
+- 1e2e787: feat: initial
+- Updated dependencies [1e2e787]
+  - @modern-js/codesmith-api-git@1.0.0-rc.13
+  - @modern-js/codesmith-api-handlebars@1.0.0-rc.13
+  - @modern-js/codesmith-api-npm@1.0.0-rc.13
+  - @modern-js/codesmith@1.0.0-rc.13
+  - @modern-js/easy-form-cli@1.0.0-rc.13
+  - @modern-js/inquirer-types@1.0.0-rc.13
+
 ## 1.0.0-rc.12
 
 ### Patch Changes