+++ packages/api/json/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @modern-js/codesmith-api-json
 
+## 1.0.0-rc.14
+
+### Patch Changes
+
+- 1e2e787: feat: initial
+- Updated dependencies [1e2e787]
+  - @modern-js/codesmith@1.0.0-rc.14
+
+## 1.0.0-rc.13
+
+### Patch Changes
+
+- 1e2e787: feat: initial
+- Updated dependencies [1e2e787]
+  - @modern-js/codesmith@1.0.0-rc.13
+
 ## 1.0.0-rc.12
 
 ### Patch Changes