+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @modern-js/codesmith-cli
 
+## 1.0.0-rc.14
+
+### Patch Changes
+
+- 1e2e787: feat: change default language key to locale
+- 1e2e787: feat: initial
+
+## 1.0.0-rc.13
+
+### Patch Changes
+
+- 1e2e787: feat: change default language key to locale
+- 1e2e787: feat: initial
+
 ## 1.0.0-rc.12
 
 ### Patch Changes