+++ packages/easy-form/cli/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @modern-js/easy-form-cli
 
+## 1.0.0-rc.14
+
+### Patch Changes
+
+- 1e2e787: feat: initial
+- Updated dependencies [1e2e787]
+  - @modern-js/easy-form-core@1.0.0-rc.14
+
+## 1.0.0-rc.13
+
+### Patch Changes
+
+- 1e2e787: feat: initial
+- Updated dependencies [1e2e787]
+  - @modern-js/easy-form-core@1.0.0-rc.13
+
 ## 1.0.0-rc.12
 
 ### Patch Changes